Fix copy_from_query()
copy_from_query() can fail when a coroutine is used as output. If the
underlying protocol connected to the backend sends more than one
CopyData message without waiting for the output coroutine, no new waiter
will have been created by the read loop in BaseProtocol.copy_out(). As a
result, waiter.cancelled() in the next call to _dispatch_result() fails.

As a quick fix, BaseProtocol now maintains a ready_for_copy_out event
which the read loop sets whenever it has created a new waiter for the
next message. This event is awaited in data_received() when the state
indicates that a COPY operation is in progress.